# Create custom image for cassandra 3.11.0