--- a/shotover-proxy/example-configs/docker-images/cassandra-3.11.13/Dockerfile
+++ b/shotover-proxy/example-configs/docker-images/cassandra-3.11.13/Dockerfile
@@ -1,0 +1,3 @@
+FROM library/cassandra:3.11.13
+
+COPY ./cassandra.yaml /etc/cassandra/cassandra.yaml