… shotover-proxy/example-configs/cassandra-cluster/docker-compose-cassandra-v3.yml
… shotover-proxy/example-configs/docker-images/cassandra-3.11.13/cassandra.yaml
… shotover-proxy/tests/test-configs/cassandra-peers-rewrite/docker-compose-3.11-cassandra.yaml