Cuts off 20s from cassandra_int_tests runtime. (10s from 2 tests)

It seems cassandra 3 is significantly faster to initialize than cassandra 4 is, so we dont save nearly as much time here as we did with the cassandra 4 PRs

With this PR they both take about 15s to start a cluster.
Before all of these speedup PRs it took:
* 25s for cassandra 3
* 70s for cassandra 4